app/requirements.txt
alembic==1.5.5
APScheduler==3.7.0
blinker==1.4
cffi==1.14.5
click==7.1.2
coverage==5.5
cryptography==3.1.1
dnspython==2.1.0
dominate==2.6.0
easy-colors==0.0.1
email-validator==1.1.2
face-recognition-models==0.3.0
flash==1.0.3
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
numpy==1.20.2
pandas==1.2.4
passlib==1.7.4
Pillow==8.1.2
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
send-email==201212
send-mail==0.0.1
six==1.15.0
SQLAlchemy==1.3.23
tzlocal==2.1
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
